refactor: shared build script to eliminate duplication
Extract common build logic into packages/shared/scripts/build.mjs
Reduces ~184 lines of duplicated code across 6 packages
Each package now uses a thin wrapper around shared script
Maintains 100% backward compatibility

Closes #482
