+++ packages/search/build.mjs
@@ -1,92 +1,16 @@
-import * as esbuild from 'esbuild'
-import { glob } from 'glob'
-import { readFileSync, writeFileSync, existsSync, mkdirSync, copyFileSync } from 'node:fs'
-import { dirname, join, relative } from 'node:path'
+#!/usr/bin/env node
 import { fileURLToPath } from 'node:url'
+import { dirname, join } from 'node:path'
+import { spawnSync } from 'node:child_process'
 
-const __dirname = dirname(fileURLToPath(import.meta.url))
+const __filename = fileURLToPath(import.meta.url)
+const __dirname = dirname(__filename)
 
-const entryPoints = await glob('src/**/*.{ts,tsx}', {
-  ignore: ['**/__tests__/**', '**/*.test.ts', '**/*.test.tsx']
+const sharedBuild = join(__dirname, '../../shared/scripts/build.mjs')
+
+const result = spawnSync('node', [sharedBuild, 'search'], {
+  stdio: 'inherit',
+  cwd: process.cwd()
 })
 
-// Plugin to add .js extension to relative imports
-const addJsExtension = {
-  name: 'add-js-extension',
-  setup(build) {
-    build.onEnd(async (result) => {
-      if (result.errors.length > 0) return
-      const outputFiles = await glob('dist/**/*.js')
-      for (const file of outputFiles) {
-        const fileDir = dirname(file)
-        let content = readFileSync(file, 'utf-8')
-        // Add .js to relative imports that don't have an extension
-        content = content.replace(
-          /from\s+["'](\.[^"']+)["']/g,
-          (match, path) => {
-            if (path.endsWith('.js') || path.endsWith('.json')) return match
-            // Check if it's a directory with index.js
-            const resolvedPath = join(fileDir, path)
-            if (existsSync(resolvedPath) && existsSync(join(resolvedPath, 'index.js'))) {
-              return `from "${path}/index.js"`
-            }
-            return `from "${path}.js"`
-          }
-        )
-        content = content.replace(
-          /import\s*\(\s*["'](\.[^"']+)["']\s*\)/g,
-          (match, path) => {
-            if (path.endsWith('.js') || path.endsWith('.json')) return match
-            // Check if it's a directory with index.js
-            const resolvedPath = join(fileDir, path)
-            if (existsSync(resolvedPath) && existsSync(join(resolvedPath, 'index.js'))) {
-              return `import("${path}/index.js")`
-            }
-            return `import("${path}.js")`
-          }
-        )
-        // Handle side-effect imports: import "./path" (no from clause)
-        content = content.replace(
-          /import\s+["'](\.[^"']+)["'];/g,
-          (match, path) => {
-            if (path.endsWith('.js') || path.endsWith('.json')) return match
-            // Check if it's a directory with index.js
-            const resolvedPath = join(fileDir, path)
-            if (existsSync(resolvedPath) && existsSync(join(resolvedPath, 'index.js'))) {
-              return `import "${path}/index.js";`
-            }
-            return `import "${path}.js";`
-          }
-        )
-        writeFileSync(file, content)
-      }
-    })
-  }
-}
-
-const outdir = join(__dirname, 'dist')
-
-await esbuild.build({
-  entryPoints,
-  outdir,
-  outbase: join(__dirname, 'src'),
-  format: 'esm',
-  platform: 'node',
-  target: 'node18',
-  sourcemap: true,
-  jsx: 'automatic',
-  plugins: [addJsExtension],
-})
-
-// Copy JSON files from src to dist (esbuild doesn't handle non-entry JSON files)
-const jsonFiles = await glob(join(__dirname, 'src/**/*.json'), {
-  ignore: ['**/node_modules/**']
-})
-for (const jsonFile of jsonFiles) {
-  const relativePath = relative(join(__dirname, 'src'), jsonFile)
-  const destPath = join(outdir, relativePath)
-  mkdirSync(dirname(destPath), { recursive: true })
-  copyFileSync(jsonFile, destPath)
-}
-
-console.log('search built successfully')
+process.exit(result.status || 0)
